+++ services/composite-keys-manager.js
@@ -8,6 +8,12 @@ function CompositeKeysManager(model, schema, record) {
     var where = {};
     var primaryKeyValues = recordId.split(GLUE);
 
+    // NOTICE: Prevent liana to crash when a composite primary keys is null,
+    //         this behaviour should be avoid instead of fixed. 
+    primaryKeyValues.forEach(function (key, index) {
+      if (key === 'null') { primaryKeyValues[index] = null; }
+    });
+
     if (primaryKeyValues.length === _.keys(model.primaryKeys).length) {
       _.keys(model.primaryKeys).forEach(function (primaryKey, index) {
         where[primaryKey] = primaryKeyValues[index];
@@ -20,6 +26,11 @@ function CompositeKeysManager(model, schema, record) {
     var compositePrimary = '';
 
     _.keys(model.primaryKeys).forEach(function (primaryKey, index) {
+      // NOTICE: Prevent liana to crash when a composite primary keys is null,
+      //         this behaviour should be avoid instead of fixed.
+      if (record[primaryKey] === null) {
+        record[primaryKey] = 'null';
+      }
       if (index === 0) {
         compositePrimary = record[primaryKey];
       } else {